+++ firmware/src/control/control_iface.h
@@ -32,10 +32,10 @@ bool control_iface_get_channel(uint channel, pwm_driver_state_t *state);
  * @brief Apply both frequency and duty to one logical channel.
  * @param channel Logical channel index.
  * @param freq_hz Requested frequency in Hz.
- * @param duty Requested duty in the normalized range `[0.0, 1.0]`.
+ * @param duty Requested duty in percent in the range `[0, 100]`; values above `100` are clamped.
  * @return Result code from the shared PWM control plane.
  */
-pwm_driver_result_t control_iface_set_channel(uint channel, float freq_hz, float duty);
+pwm_driver_result_t control_iface_set_channel(uint channel, uint32_t freq_hz, uint8_t duty);
 
 /**
  * @brief Apply frequency while preserving the currently realized duty.
@@ -43,15 +43,15 @@ pwm_driver_result_t control_iface_set_channel(uint channel, float freq_hz, float
  * @param freq_hz Requested frequency in Hz.
  * @return Result code from the shared PWM control plane.
  */
-pwm_driver_result_t control_iface_set_channel_freq(uint channel, float freq_hz);
+pwm_driver_result_t control_iface_set_channel_freq(uint channel, uint32_t freq_hz);
 
 /**
  * @brief Apply duty while preserving the currently realized frequency.
  * @param channel Logical channel index.
- * @param duty Requested duty in the normalized range `[0.0, 1.0]`.
+ * @param duty Requested duty in percent in the range `[0, 100]`; values above `100` are clamped.
  * @return Result code from the shared PWM control plane.
  */
-pwm_driver_result_t control_iface_set_channel_duty(uint channel, float duty);
+pwm_driver_result_t control_iface_set_channel_duty(uint channel, uint8_t duty);
 
 /**
  * @brief Stop all channels and restore the shared default state.
